requirements.txt
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
Django~=4.0.3
django-cors-headers==3.12.0
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
executing==0.8.3
factory-boy==3.2.1
Faker==13.3.0
iniconfig==1.1.1
ipython
jedi==0.18.1
matplotlib-inline==0.1.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.28
pure-eval==0.2.2
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
tzdata==2021.5
wcwidth==0.2.5
